Provide a dependency-less `BloggingPromptsAttribution` definition
This will allow us to move `BloggingPromptsAttribution` to WordPressData as
part of #24165 without dragging over those dependencies.
